requirements.txt
streamlit==1.32.0
opencv-python==4.8.1.78
numpy==1.26.4
llama-index==0.9.48
pillow==10.2.0

